+++ ChangeLog.md
@@ -6,6 +6,9 @@
 
     Release date: xxx
 
+    - pwiAccounts: add new 'onVerifiedEmailBox' configuration parameter
+    - pwiAccounts: add new 'onVerifiedEmailTitle' configuration parameter
+    - pwiAccounts: add new 'onVerifiedEmailMessage' configuration parameter
     - pwiAccounts: add new 'sendVerificationEmail' configuration parameter
     - acUserLogin: add new 'haveCancelButton' configuration parameter